+++ src/client.jl
@@ -1,6 +1,12 @@
 # Constants
 # ---------
-const HEADERS = Dict("Authorization" => "DeepL-Auth-Key " * DEEPL_API_KEY, "Content-Type" => "application/json")
+const AUTH_HEADERS = Dict(
+    "Authorization" => "DeepL-Auth-Key " * DEEPL_API_KEY
+)
+const HEADERS = Dict(
+    "Authorization" => "DeepL-Auth-Key " * DEEPL_API_KEY, 
+    "Content-Type" => "application/json"
+)
 
 """
     post_request(endpoint::String, body::Dict)
@@ -15,6 +21,22 @@ Make a POST request to the DeepL API.
 function post_request(endpoint::String, body::Dict)
     url = DEEPL_API_URL * endpoint
     response = HTTP.post(url, HEADERS, JSON.json(body))
+    return response
+end
+
+"""
+    get_request(endpoint::String; params::Dict=Dict())
+
+Make a GET request to the DeepL API.
+# Arguments
+- `endpoint::String`: The endpoint to request.
+- `params::Dict`: Optional query parameters.
+# Returns
+- `HTTP.Response`: The HTTP response from the API.
+"""
+function get_request(endpoint::String; params::Dict=Dict())
+    url = DEEPL_API_URL * endpoint
+    response = HTTP.get(url, AUTH_HEADERS; query=params)
     return response
 end